# openapi.yaml
openapi: 3.1.0
info:
  title: Jitsi Portal API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/me:
    get:
      summary: Get current authenticated user profile
      operationId: getAuthMe
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SafeUserProfileResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
  /api/v1/auth/refresh:
    post:
      summary: Refresh access token
      operationId: refreshAccessToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRequest'
      responses:
        '200':
          description: Token pair refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRefreshResponse'
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Token revoked
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '409':
          description: Refresh token reuse detected
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
  /api/v1/auth/refresh/revoke:
    post:
      summary: Revoke refresh token by token id
      operationId: revokeRefreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRefreshRevokeRequest'
      responses:
        '204':
          description: Token revoked
        '400':
          description: Invalid request payload
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '401':
          description: Authentication required
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
        '403':
          description: Access denied
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetailResponse'
...
components:
  schemas:
...
    SafeUserProfileResponse:
      type: object
      properties:
        id:
          type: string
        displayName:
          type: string
        email:
          type: string
        tenant:
          type: string
        claims:
          type: object
          additionalProperties:
            type: string
      required:
        - id
        - displayName
        - email
        - tenant
        - claims
    AuthRefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
      required:
        - refreshToken
    AuthRefreshResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresAt:
          type: string
          format: date-time
        role:
          type: string
          enum: [participant, moderator, host]
        tokenType:
          type: string
          enum: [Bearer]
      required:
        - accessToken
        - refreshToken
        - expiresAt
        - role
        - tokenType
    AuthRefreshRevokeRequest:
      type: object
      properties:
        tokenId:
          type: string
      required:
        - tokenId
...
    ProblemDetailResponse:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        properties:
          type: object
          properties:
            errorCode:
              type: string
            traceId:
              type: string
          required:
            - errorCode
            - traceId
...